Fix: - Bug with adding the same product, need to update after change quantity of product in order
Chore:
- Adding product to order via email instead id
- Delete product
- Buying Order
- Cant buy empty order
- Cant delete product from buying order

+++ src/main/kotlin/com/andruszkiewiczartur/Internet/shop/serwice/impl/QuantityServiceImpl.kt
@@ -1,9 +1,11 @@
 package com.andruszkiewiczartur.Internet.shop.serwice.impl
 
+import com.andruszkiewiczartur.Internet.shop.domain.entity.OrderEntity
 import com.andruszkiewiczartur.Internet.shop.domain.entity.QuantityEntity
 import com.andruszkiewiczartur.Internet.shop.repository.OrderRepository
 import com.andruszkiewiczartur.Internet.shop.repository.ProductRepository
 import com.andruszkiewiczartur.Internet.shop.repository.QuantityRepository
+import com.andruszkiewiczartur.Internet.shop.repository.UserRepository
 import com.andruszkiewiczartur.Internet.shop.serwice.QuantityService
 import org.springframework.stereotype.Service
 
@@ -11,18 +13,84 @@ import org.springframework.stereotype.Service
 class QuantityServiceImpl(
     private val quantityRepository: QuantityRepository,
     private val orderRepository: OrderRepository,
-    private val productRepository: ProductRepository
+    private val productRepository: ProductRepository,
+    private val userRepository: UserRepository
 ): QuantityService {
 
-    override fun addProduct(orderId: Long, productId: Long, quantity: Int): QuantityEntity =
-        quantityRepository.save(
-            QuantityEntity(
-                order = orderRepository.getReferenceById(orderId),
-                product = productRepository.getReferenceById(productId),
-                quantity = quantity
-            )
-        )
+    override fun addProduct(email: String, productId: Long, quantity: Int): QuantityEntity {
+        val user = userRepository
+            .getReferenceById(email)
 
-    override fun deleteProduct(quantityId: Long) =
-        quantityRepository.deleteById(quantityId)
+        var order = user
+            .orders
+            .find { !it.isBuying }
+
+        if (order == null) {
+            order = orderRepository.save(
+                OrderEntity(
+                    products = mutableListOf(),
+                    isBuying = false
+                )
+            )
+
+            userRepository.save(
+                user.copy(
+                    orders = user.orders.toMutableList().also { it.add(order) }
+                )
+            )
+        }
+
+        val orderProduct = order.products.find { it.product.id == productId }
+
+        val quantityResponse = if (orderProduct == null) {
+            val product = productRepository.findById(productId).get()
+
+            quantityRepository.save(
+                QuantityEntity(
+                    product = product,
+                    quantity = quantity
+                )
+            )
+        } else {
+            quantityRepository.save(
+                orderProduct.copy(
+                    quantity = quantity
+                )
+            )
+        }
+
+        val newOrder = if (quantityResponse != orderProduct?.product) {
+            orderRepository.save(
+                order.copy(
+                    products = order.products.toMutableList().also {
+                        it.addLast(quantityResponse)
+                    }
+                )
+            )
+        } else {
+            order
+        }
+
+        return newOrder.products.last()
+    }
+
+    override fun deleteProduct(quantityId: Long) {
+        val order = orderRepository
+            .findAll()
+            .firstOrNull { it.products.firstOrNull { it.id == quantityId } != null }
+            ?: throw RuntimeException()
+
+        if(!order.isBuying) {
+            val product = order.products.first { it.id == quantityId }
+
+            orderRepository.save(
+                order.copy(
+                    products = order.products.toMutableList().also {
+                        it.remove(product)
+                    }
+                )
+            )
+            quantityRepository.deleteById(quantityId)
+        }
+    }
 }
